task 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -21,8 +21,10 @@ def delete_state_a():
 
     session = Session(engine)
 
-    states_to_delete = session.query(State).filter(state.name.like(%a%)).all()
-    for state in states_to_delete:
+    states = session.query(State).all()
+
+    for state in states:
+        if 'a' in state.name:
             session.delete(state)
 
     session.commit()
